+++ src/index.ts
@@ -1,3 +1,3 @@
-export { resolveJSONSchema, generateJoi, formatJoi, Options, SubSchemas } from './joi';
+export { resolveJSONSchema, generateJoiStatement, formatJoi, Options, SubSchemas } from './joi';
 import * as Joi from './extendedJoi';
 export { Joi };